+++ relay-dynamic-config/src/feature.rs
@@ -118,8 +118,10 @@ pub enum Feature {
     #[serde(rename = "projects:span-v2-experimental-processing")]
     SpanV2ExperimentalProcessing,
     /// Enables OTLP spans to use the Span V2 processing pipeline in Relay.
+    ///
+    /// This is now the default behaviour of Relay.
     #[serde(rename = "organizations:span-v2-otlp-processing")]
-    SpanV2OtlpProcessing,
+    DeprecatedSpanV2OtlpProcessing,
     /// This feature has deprecated and is kept for external Relays.
     #[doc(hidden)]
     #[serde(rename = "projects:span-metrics-extraction")]